Allow dependency confusion mitigation by grouping ecosystems
Currently all ecosystems are given equal considerations when making
a recommendation for a namespace. This change changes the array of
recommendation managers to an array of an array of recommendation
managers. This allows a setup like `[[Darkpan],[fez,p6c]]` where
if `Foo` is found in Darkpan it will always be chosen, even if fez
or p6c have a `Foo` with a higher version.

Note this is incomplete -- some of the deduplicatin logic currently
in Zef::Client!find-prereq-candidates needs to be moved to Zef::Repository
so if it e.g. only finds 1 of 2 namespaces it queried for that it
searches the next group (and only for the missing namespace).